test: move export_config fixture to apps/exports/tests/fixtures.py
export_config lived inline in test_list_view.py; relocate it to the shared
exports fixtures module so other test modules (and the multi-project / run
fixtures added next) can reuse it via import.
